Refactor. gsap 이벤트 class 객체
