+++ src/containers/home/works/WorksAnimationContainer.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Work } from "@/services/works.service";
-import { MainWorksListEffect } from "@/utils/lib/gsap";
+import Animate from "@/utils/lib/gsap";
 import { scrollOffsetYState } from "@/utils/lib/recoil/atom";
 import Image from "next/image";
 import { useEffect, useRef } from "react";
@@ -17,6 +17,7 @@ export default function WorksAnimationContainer({ works }: Props) {
 
   useEffect(() => {
     if (!containerRef.current) return;
+    const { MainWorksListEffect } = new Animate();
     const container = containerRef.current;
     const containerTop = container.offsetTop;
